solved windows warnings

+++ sdk/src/connections/usb/windows/usb_storage_windows.cpp
@@ -81,7 +81,7 @@ Status UsbStorage::read(const uint32_t address, uint8_t *data,
 
     HRESULT hr = UsbWindowsUtils::UvcExUnitReadBuffer(
         m_implData->handle->pVideoInputFilter, 5, m_implData->id, address, data,
-        bytesCount);
+        static_cast<uint32_t>(bytesCount));
     if (FAILED(hr)) {
         LOG(WARNING)
             << "Failed to read buffer through UVC extension unit. Error: "
@@ -104,7 +104,7 @@ Status UsbStorage::write(const uint32_t address, const uint8_t *data,
 
     HRESULT hr = UsbWindowsUtils::UvcExUnitWriteBuffer(
         m_implData->handle->pVideoInputFilter, 6, m_implData->id, address, data,
-        bytesCount);
+        static_cast<uint32_t>(bytesCount));
     if (FAILED(hr)) {
         LOG(WARNING)
             << "Failed to write buffer through UVC extension unit. Error: "
